adapter: check wake_override when setting device privacy
For the device using a RPA, hog_probe sets wake_override to true, but
the command to set remote wakeup fails because the device has not been
added to the kernel and the connection with the public address cannot be
found.

When setting the device privacy flag, the wakeup flag should be set
according to the wake_override, instead of the current flags.

Signed-off-by: Zhengping Jiang <jiangzp@google.com>