adapter: Enable SSP after enabling BREDR
When enabling Secure Simple Pairing, the kernel first checks if
BREDR is enabled. If it's not, enabling SSP fails. Therefore, it
is necessary to make sure that BREDR is enabled before enabling
SSP.

Without this fix, bluetoothd would start without SSP enabled if
started in dual or bredr mode after it had been run in le mode.