plugins: convert external sixaxis plugin to builtin
Sixaxis plugin is the only external one. It's a tiny 20K binary that
distros ship a separate package for.

Make it a builtin, which allows distros to drop the separate package, it
also enables us to remove support for external modules - both in terms
of extra code and hide the internal bluetoothd API.

This means that libudev.so is pulled in, which is fine since its ABI has
been stable for over a decade.