shared/bap: Check state of all streams with same BIG ID
The function is used to verify if all the streams from a BIG are in the
same state, and if so it returns a new queue with all these streams.
This queue will be used to iterate through all configured streams in order
of their creation, and create the IOs.