build: handle relative libexecdir instances
With earlier commit, we've used exec_dir to reference the path of obexd
and bluetoothd within the service files.

At the same time, in my testing I was providing the complete path on the
configure command line. As result, things just worked.

In the default case, the variable contains relative references to ${}
variables, which as result end up literal in the services. Ultimately
the service files were broken.

Document are reuse the existing pattern of manually expanding the
variables.

Closes: https://github.com/bluez/bluez/issues/736

Cc: Joakim Tjernlund <Joakim.Tjernlund@infinera.com>