profiles: Add initial code for an ASHA plugin
This implements the server role for the Audio Streaming for Hearing Aid
specification[1].

The implementation registers a remote endpoint using a subset of the
MediaEndpoint1 interface, without any mechanism for setting/selecting a
configuration, as this is all static in the spec for now. Also exposed
on connection is a MediaTransport1 object, which can be used to obtain
an fd to stream to the device.

[1] https://source.android.com/docs/core/connect/bluetooth/asha