bap: Create transports for matching BISes
Rename bt_bap_add_bis to bt_bap_check_bis. This function is now an
utilitary for checking if the information from a BIS is matching the
capabilities in bt_bap. This utilitary is used when parsing the BASE
data so that we can decide if we create a new setup/stream/transport for
it, or advance to the next BIS.
When checking in local PACS for a BIS match, return also the local PAC,
along with the status of the operation. This information is required
later when we create the stream.

Use the BIS index received in the BASE to synchronize to the BIG.
Allow bt_bap_endpoint to be NULL.
Remove the Broadcast Sink code from set_configuration.
Update BASE parsing so that it creates streams and transports, without a
remote PAC and endpoint.
Update bap_find_setup_by_stream to find the setup in case the stream does
not contain an endpoint.