set: Attempt to use existing set gatt-db
Most sets should be clone of each other, or at least very similar, so
this attempts to clone the existing gatt-db of the first member found
when connecting new sets, this substantially speed up the process of
bonding sets if their database matches which is something that is
currently ranging from 20-30 seconds depending on the manufacturer and
with this changes it takes a good 5-10 seconds.

If the dbs don't really match bt_gatt_client instance will attempt to
rediscover the ranges that don't match.