doc: Add initial MediaSetup rst
This proposes a new D-Bus API to be used by the BAP Broadcast Assistant
role.

The BAP Broadcast Assistant is a Master device that can have multiple
connections with BAP Scan Delegators. The Broadcast Assistant has the UI
to start scanning, enumerate Broadacast Sources, allow the user to select
a desired Source and send the stream information to one of the peers. If
the selected stream is encrypted and the peer requires a Broadcast Code
to decrypt the stream, the user should issue a command to send the code.

A new D-Bus API is necessary to allow the BAP Broadcast Assistant
implementation in BlueZ to communicate with a top level UI process.

The Broadcast Assistant in BlueZ should register a D-Bus object for each
scanned BIS that matches the Audio capabilities of a peer. Each object
should be linked with the stream and the peer device that matches the
stream capabilities.

The top level process should detect these objects and display the
information to the user. The user can issue commands to select an object
or to provide the Broadcast Code for an object.

More information on the proposed design can be found here:
https://github.com/orgs/bluez/discussions/822