transport: Broadcast sink: wait for user to select transport
This changes the flow for transports created on broadcast sink side.
Transports are not automatically changed to pending anymore, instead
the user must first run transport.select on them. This allows for the
selection of the desired stream when running the setup with PipeWire,
which acquires any transport that is pending.