gdbus: Add g_dbus_proxy_set_property_dict
This adds a new gdbus function to allow for the setting of a
dictionary type property - a{sv}. It received the name of the
property, the name of the entry to be set from the dictionary,
it's type and value. These values are passed as a dict entry,
to the set method associated with the property. There, it must
be parsed and the appropriate entry updated.