Client: Fix the list_attributes command returning nothing for a dual-mode remote
When a dual-mode device is paired first on BR/EDR and
then on BLE through RPA, the RPA changes to a public
address after receiving the IRK. This results in two proxies
in default_ctrl->devices with the same public address.
In cmd_list_attributes, if the BR/EDR proxy is found first,
it prints no attributes.

Modify cmd_list_attributes to search all proxies in
default_ctrl->devices and display the related attributes.